Fixed and unified blank spaces
There was a bug that didn't allow the items of lists in the
documentation to be separated by one or more blank lines.

Now blank lines are processed so that they leave an actual blank line in
the html code, instead of using the <p> tag.

This improves the looks of the result.